+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -7,7 +7,7 @@ import java.util.Collection;
 public interface UserStorage {
     User addUser(User user);
     User updateUser(User user);
-    void deleteUser(Integer id);
-    User findUserById(Integer id);
+    void deleteUser(Long id);
+    User findUserById(Long id);
     Collection<User> findAllUsers();
 }
